requirements.txt
behave>=1.2.1
Django==1.4.1
dj-database-url==0.2.1
gunicorn==0.14.6
psycopg2==2.4.5
python-dateutil==2.1
requests==0.13.5
